Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==3.0.0

SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

flask-jwt-extended==4.5.0

requests==2.31.0

structlog==23.2.0

pytest==7.4.3
